feat: cambio en la vista de ordenes para ver por supplier
